Release client-go 1.0

See merge request gitlab-org/api/client-go!2575

BREAKING CHANGE: Release 1.0

+++ commits_test.go
@@ -102,7 +102,7 @@ func TestGetCommitStatuses_WithOptions(t *testing.T) {
 		Ref:        Ptr("master"),
 		Stage:      Ptr("test"),
 		Name:       Ptr("ci"),
-		PipelineID: Ptr(1),
+		PipelineID: Ptr(int64(1)),
 		All:        Ptr(true),
 	}
