+++ job_token_scope_test.go
@@ -146,7 +146,7 @@ func TestAddProjectToJobScopeAllowList(t *testing.T) {
 
 	addTokenResponse, resp, err := client.JobTokenScope.AddProjectToJobScopeAllowList(
 		1,
-		&JobTokenInboundAllowOptions{TargetProjectID: Ptr(2)},
+		&JobTokenInboundAllowOptions{TargetProjectID: Ptr(int64(2))},
 	)
 	assert.NoError(t, err)
 	assert.Equal(t, want, addTokenResponse)
@@ -246,7 +246,7 @@ func TestAddGroupToJobTokenAllowlist(t *testing.T) {
 
 	addTokenResponse, resp, err := client.JobTokenScope.AddGroupToJobTokenAllowlist(
 		1,
-		&AddGroupToJobTokenAllowlistOptions{TargetGroupID: Ptr(2)},
+		&AddGroupToJobTokenAllowlistOptions{TargetGroupID: Ptr(int64(2))},
 	)
 	assert.NoError(t, err)
 	assert.Equal(t, want, addTokenResponse)
